added token, login

+++ controllers/cards.js
@@ -1,4 +1,7 @@
+const escape = require('escape-html');
+
 const Cards = require('../models/card');
+const User = require('../models/user');
 const { handleEmptyData } = require('./services/dataHandlers');
 const { checkIdValiness } = require('./services/checkIdValiness');
 const { handleNotFound } = require('./services/handleNotFound');
@@ -13,7 +16,11 @@ module.exports.readCards = (req, res) => {
 module.exports.createCard = (req, res) => {
   const { name, link } = req.body;
 
-  Cards.create({ name, link, owner: req.user._id })
+  Cards.create({
+    name: escape(name),
+    link: escape(link),
+    owner: req.user._id,
+  })
     .then((card) => res.send({ data: card }))
     .catch((err) => res.status(500).send({ message: `Произошла ошибка, ${err}` }));
 };
@@ -21,14 +28,24 @@ module.exports.createCard = (req, res) => {
 module.exports.deleteCard = async (req, res) => {
   const { cardId } = req.params;
   const isIdValid = await checkIdValiness(Cards, cardId);
+  const { _id } = req.user;
 
-  if (isIdValid) {
+  let isUserCardOwner = false;
+  User.findById(_id)
+    .then((user) => {
+      isUserCardOwner = _id === user._id;
+    })
+    .catch(() => res.status(403).sned({ message: 'действие недоступно пользователю' }));
+
+  if (isIdValid && isUserCardOwner) {
     Cards.findByIdAndRemove(cardId)
       .then((card) => handleEmptyData(card, res))
       .then(() => res.send({ message: 'данные обновлены' }))
       .catch((err) => res.status(500).send({ message: `Произошла ошибка, ${err}` }));
-  } else {
+  } else if (!isIdValid) {
     handleNotFound(res);
+  } else if (!isUserCardOwner) {
+    res.status(403).sned({ message: 'действие недоступно пользователю' });
   }
 };
